requirements.txt
streamlit==1.37.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.36
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
openpyxl==3.1.2
